goabackend: add GoaServiceAuth and GoaPasswordAuth
Add a base class for credentials, similar to GoaServiceConfig.
Included is a basic implementation for password-cleartext and
password-encrypted (as known in Mail Autoconfig).

The authentication method, like the config class, is a string
with constants defined from the Mail Autoconfig specification. The
password auth may be used directly or subclassed, as the service
client is responsible for performing authentication.