goadaemon: Delete groups without provider in add_config_file()
It could sometimes happen that the config file contains incomplete
sections, which have missing the 'Provider' key. They are skipped
in the code, but they are left in the file, thus whenever the file
is read they are skipped again and a warning is printed.

This change will clean up the config file from any such leftovers.