goaoauth2provider: rewrite the D-Bus helper as a per-task service
Rewrite the authorization task so the D-Bus service is tied to
the lifetime of the GTask object. When the task completes, the
service is shutdown and while it runs, no other can displace it.

Usually only one process (e.g. GNOME Settings) performs a request
at a time by using a modal dialog, but this will prevent others
from accidentally doing so and throw an error if they try, thus
preventing any intentional or unintentional denial-of-service or
race conditions.